refactor(harmonizeSource): remove HTML-specific extraction; keep shared utilities
Deleted extractValues, setNestedProperty, removeTrailingSlash,
MAX_SELECTOR_LENGTH/DEPTH/RULES, and the harmonizeSource export function.
Updated public exports to expose mergeSchemas, processValue, filterValues
alongside remoteHarmonizer. Updated src/lib/indexing.js to drop the dead
harmonizeSource dep from createIndexer. Updated test files to call the HTML
handler's harmonize() directly instead of the removed harmonizeSource export.
Also fixed a handler null-guard bug (undefined harmonizerSchema was erroneously
entering the string branch).
