+++ src/tests/exports.test.js
@@ -8,12 +8,12 @@ describe('octothorpes package exports', () => {
     'buildMultiPass', 'getBlobjectFromResponse', 'createHarmonizerRegistry',
     'parseUri', 'validateSameOrigin', 'getScheme',
     'verifiedOrigin', 'parseBindings', 'deslash', 'getFuzzyTags', 'isSparqlSafe',
-    'rss', 'arrayify', 'harmonizeSource',
+    'rss', 'arrayify',
     'createIndexer', 'resolveSubtype', 'isHarmonizerAllowed',
     'checkIndexingRateLimit', 'checkIndexingPolicy', 'parseRequestBody', 'isURL',
     // newly added
     'badgeVariant', 'determineBadgeUri',
-    'remoteHarmonizer',
+    'remoteHarmonizer', 'mergeSchemas', 'processValue', 'filterValues',
     'verifyApprovedDomain',
     'createEnrichBlobjectTargets',
     // utils additions
